--- a/src/constructors/css.js
+++ b/src/constructors/css.js
@@ -13,5 +13,6 @@ export default (
   ) {
     return flatten(interleave([], [styles, ...interpolations]))
   }
+
   return flatten(interleave(styles, interpolations))
 }