--- a/src/constructors/css.js
+++ b/src/constructors/css.js
@@ -1,5 +1,7 @@
 // @flow
 import interleave from '../utils/interleave'
+import isPlainObject from '../utils/isPlainObject'
+import { EMPTY_ARRAY } from '../utils/empties'
 import flatten from '../utils/flatten'
 import type { Interpolation, RuleSet, Styles } from '../types'
 
@@ -7,12 +9,11 @@ export default (
   styles: Styles,
   ...interpolations: Array<Interpolation>
 ): RuleSet => {
-  if (
-    !Array.isArray(styles) &&
-    (typeof styles === 'object' || typeof styles === 'function')
-  ) {
-    return flatten(interleave([], [styles, ...interpolations]))
+  if (typeof styles === 'function' || isPlainObject(styles)) {
+    // $FlowFixMe
+    return flatten(interleave(EMPTY_ARRAY, [styles, ...interpolations]))
   }
 
+  // $FlowFixMe
   return flatten(interleave(styles, interpolations))
 }