--- a/src/hoc/withTheme.js
+++ b/src/hoc/withTheme.js
@@ -1,18 +1,13 @@
 // @flow
 import React, { type ComponentType } from 'react'
-import PropTypes from 'prop-types'
 import hoistStatics from 'hoist-non-react-statics'
-import {
-  CHANNEL,
-  CHANNEL_NEXT,
-  CONTEXT_CHANNEL_SHAPE,
-} from '../models/ThemeProvider'
+import { CHANNEL_NEXT, contextShape } from '../models/ThemeProvider'
 import { EMPTY_OBJECT } from '../utils/empties'
+import getComponentName from '../utils/getComponentName'
 import _isStyledComponent from '../utils/isStyledComponent'
 import determineTheme from '../utils/determineTheme'
 
-const wrapWithTheme = (Component: ComponentType<any>) => {
-  const componentName = Component.displayName || Component.name || 'Component'
+export default (Component: ComponentType<any>) => {
   const isStatelessFunctionalComponent =
     typeof Component === 'function' &&
     // $FlowFixMe TODO: flow for prototype
@@ -23,16 +18,12 @@ const wrapWithTheme = (Component: ComponentType<any>) => {
     _isStyledComponent(Component) || isStatelessFunctionalComponent
 
   class WithTheme extends React.Component<*, *> {
-    static displayName = `WithTheme(${componentName})`
+    static contextTypes = contextShape
+    static displayName = `WithTheme(${getComponentName(Component)})`
     static defaultProps: Object
 
     // NOTE: This is so that isStyledComponent passes for the innerRef unwrapping
     static styledComponentId = 'withTheme'
-
-    static contextTypes = {
-      [CHANNEL]: PropTypes.func,
-      [CHANNEL_NEXT]: CONTEXT_CHANNEL_SHAPE,
-    }
 
     state: { theme?: ?Object } = EMPTY_OBJECT
     unsubscribeId: number = -1
@@ -90,11 +81,9 @@ const wrapWithTheme = (Component: ComponentType<any>) => {
         delete props.innerRef
       }
 
-      return <Component {...props} />
+      return React.createElement(Component, props)
     }
   }
 
   return hoistStatics(WithTheme, Component)
 }
-
-export default wrapWithTheme