--- a/src/models/ComponentStyle.js
+++ b/src/models/ComponentStyle.js
@@ -9,7 +9,7 @@ import isStyledComponent from '../utils/isStyledComponent'
 const areStylesCacheable = IS_BROWSER
 
 const isStaticRules = (rules: RuleSet, attrs?: Object): boolean => {
-  for (let i = 0; i < rules.length; i += 1) {
+  for (let i = 0, len = rules.length; i < len; i += 1) {
     const rule = rules[i]
 
     // recursive case
@@ -25,8 +25,7 @@ const isStaticRules = (rules: RuleSet, attrs?: Object): boolean => {
   if (attrs !== undefined) {
     // eslint-disable-next-line guard-for-in, no-restricted-syntax
     for (const key in attrs) {
-      const value = attrs[key]
-      if (typeof value === 'function') {
+      if (typeof attrs[key] === 'function') {
         return false
       }
     }
@@ -91,8 +90,11 @@ export default (
       const name = generateRuleHash(this.componentId + flatCSS.join(''))
 
       if (!styleSheet.hasNameForId(componentId, name)) {
-        const css = stringifyRules(flatCSS, `.${name}`)
-        styleSheet.inject(this.componentId, css, name)
+        styleSheet.inject(
+          this.componentId,
+          stringifyRules(flatCSS, `.${name}`),
+          name
+        )
       }
 
       this.lastClassName = name