--- a/src/utils/empties.js
+++ b/src/utils/empties.js
@@ -1,0 +1,3 @@
+// @flow
+export const EMPTY_ARRAY = Object.freeze([])
+export const EMPTY_OBJECT = Object.freeze({})