--- a/src/utils/isPlainObject.js
+++ b/src/utils/isPlainObject.js
@@ -1,0 +1,2 @@
+// @flow
+export default (x: any) => typeof x === 'object' && x.constructor === Object