The broad strokes:

- use mutation instead of allocation wherever possible
- unpacking closures so the VM can cache things better (important for SSR performance)

So far:

### before

---Mount deep tree---
20.82 ±02.67 ms
(S/L) 16.45/04.37 ms

---Mount wide tree---
31.97 ±03.22 ms
(S/L) 25.45/06.51 ms

---Update dynamic styles---
06.90 ±01.10 ms
(S/L) 06.29/00.61 ms

### after

---Mount deep tree---
17.09 ±01.77 ms
(S/L) 13.01/04.08 ms

---Mount wide tree---
24.34 ±02.19 ms
(S/L) 18.25/06.09 ms

---Update dynamic styles---
05.56 ±01.08 ms
(S/L) 05.00/00.56 ms